requirements.txt
Flask==2.0.1
opencv-python-headless==4.5.3.56
Pillow==8.2.0
numpy==1.21.0